使用SSBO重写光线追踪

+++ chinese-chess/tools/image_helper.cpp
@@ -16,7 +16,7 @@ void image_helper::read_hdr_image(const std::filesystem::path& _hdr_path, uint32
 	_width = spec.width;
 	_height = spec.height;
 
-	_hdr_data.resize(_width * _height * 4, 1.f);
+	_hdr_data.resize(static_cast<size_t>(_width) * _height * 4, 1.f);
 	if (!image->read_image(0, 0, 0, 4, OIIO::TypeDesc::FLOAT, _hdr_data.data(), 4 * sizeof(float), 4 * sizeof(float) * _width, OIIO::AutoStride))
 	{
 		throw std::runtime_error(std::format("Failed to read image {}.", _hdr_path.string()));
